requirements.txt
# Phase 0 data collection (see docs/phase0.md)
openai>=1.0.0
transformers>=4.30.0
pyyaml>=6.0
requests>=2.28.0

# Kconfig support (for make menuconfig)
kconfiglib>=14.1.0

# Correctness evaluator: RemoteVerdictEvaluator lives in minions_channel (sibling repo).
# Add that repo to PYTHONPATH or set MINIONS_CHANNEL_PATH to its root when running.
